// src/app/Modules/rrhh/model/relpuepers.dto.ts
export interface RelPuePersDTO {
    activoHoy:                   string;
    administracionProcedenteCod: string;
    administracionProcedenteDes: string;
    adscripcionCod:              string;
    adscripcionDes:              string;
    aextinguirDot:               string;
    aextinguirPto:               string;
    agrupaPuestoCod:             number;
    apellido1:                   string;
    apellido2:                   string;
    apellidosNombre:             string;
    categoriaFaseCod:            number;
    categoriaFaseDesc:           string;
    centroCos1Cod:               string;
    centroCos1Des:               string;
    centroCos2Cod:               string;
    centroCos2Des:               string;
    centroCos3Cod:               string;
    centroCos3Des:               string;
    centroCos4Cod:               string;
    centroCos4Des:               string;
    centroCos5Cod:               string;
    centroCos5Des:               string;
    compatibilidad:              string;
    compatibilidadDes:           string;
    convenioCod:                 number;
    convenioDes:                 string;
    convenioFaseCod:             number;
    convenioFaseDes:             string;
    correoElectronico:           string;
    cotizacionCod:               number;
    desempenado2actividad:       string;
    desempenadoPlaboral:         string;
    destinoCod:                  number;
    destinoPtoCod:               number;
    direccionEdificioDota:       string;
    edad:                        number;
    edificioDotaCod:             number;
    empleadoCod:                 number;
    emprecotiFaseCod:            number;
    emprecotiFaseDesc:           string;
    entidadDes:                  string;
    escalaPuestoCod:             string;
    escalaPuestoDes:             string;
    especificoCod:               number;
    especificoPtoCod:            number;
    estadoCoberturaCod:          string;
    estadoCoberturaDes:          string;
    expedienteCese:              string;
    expedienteCeseFecha:         Date;
    expedienteNombr:             string;
    expedienteNombrFecha:        string;
    expedienteSitu:              string;
    expedienteSituFecha:         string;
    fechaNacimiento:             string;
    grupoFaseCod:                string;
    grupoFaseDesc:               string;
    grupoRptgenericoCod:         string;
    grupoRptocupacionCod:        string;
    id:                          RelPuePersIdDTO;
    inicioPosesionPto:           Date;
    inicioVersionDot:            Date;
    inicioVersionFas:            Date;
    inicioVersionPto:            Date;
    finPosesionPto:              Date;
    finVersionDot:               Date;
    finVersionFas:               Date;
    finVersionPto:               Date;
    jornadaReducida:             string;
    motivoAdscPtoTex:            string;
    motivoCesePto:               string;
    motivoSituPtoTex:            string;
    nif:                         string;
    nivelCod:                    string;
    nivelEspecificoDes:          string;
    nivelEspecificoPtoDes:       string;
    nombre:                      string;
    observacionesFas:            string;
    ocupacionCod:                string;
    ocupacionFaseCod:            string;
    organizacionNivel1Cod:       string;
    organizacionNivel1Des:       string;
    organizacionNivel2Cod:       string;
    organizacionNivel2Des:       string;
    organizacionNivel3Cod:       string;
    organizacionNivel3Des:       string;
    organizacionNivel4Cod:       string;
    organizacionNivel4Des:       string;
    organizacionNivel5Cod:       string;
    organizacionNivel5Des:       string;
    porcentajeJornadaDot:        number;
    porcentajeJornadaFas:        number;
    presuDiadesdeDotacion:       number;
    presuDiafinDotacion:         number;
    presuMesdesdeDotacion:       number;
    presuMesfinDotacion:         number;
    provisionCod:                string;
    provisionDes:                string;
    ptoResponsable:              string;
    puestoClave:                 string;
    puestoDes:                   string;
    regimenFaseCod:              number;
    regimenFaseDesc:             string;
    relacionEmpleoCod:           string;
    relacionEmpleoDes:           string;
    rptDes:                      string;
    sexo:                        string;
    singular:                    string;
    singularDes:                 string;
    situacionCod:                string;
    situacionDes:                string;
    tipoJornada:                 string;
    titular:                     string;
    zonaDotaCod:                 number;
}




export interface RelPuePersIdDTO {
    dotacionNum:       number;
    entidadCod:        number;
    faseNumero:        number;
    puestoCod:         number;
    rptCod:            string;
    versionFasePuesto: number;
    versionDotacion:             number;
    versionPuesto:               number;
}